+++ src/cmd/count.rs
@@ -73,18 +73,13 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
             }
         }
     } else {
-        match conf.indexed()? {
-            Some(idx) => idx.count(),
-            None => {
-                let mut rdr = conf.reader()?;
-                let mut count = 0u64;
-                let mut record = csv::ByteRecord::new();
-                while rdr.read_byte_record(&mut record)? {
-                    count += 1;
-                }
-                count
-            }
+        let mut rdr = conf.reader()?;
+        let mut count = 0u64;
+        let mut record = csv::ByteRecord::new();
+        while rdr.read_byte_record(&mut record)? {
+            count += 1;
         }
+        count
     };
 
     if args.flag_csv {