+++ src/cmd/slice.rs
@@ -1,8 +1,6 @@
-use std::fs;
 use std::io::{Read, SeekFrom};
 
 use crate::config::{Config, Delimiter};
-use crate::index::Indexed;
 use crate::util;
 use crate::CliResult;
 
@@ -15,11 +13,6 @@ record in the CSV data.
 
 If the end of the range isn't specified, then the slice continues to the last
 record in the CSV data.
-
-This operation can be made much faster by creating an index with 'xan index'
-first. Namely, a slice on an index requires parsing just the rows that are
-sliced. Without an index, all rows up to the first row in the slice must be
-parsed.
 
 Finally, this command is also able to find the first record to slice in
 constant time using the -B, --byte-offset if you know its byte offset in
@@ -83,51 +76,43 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
 
     match &args.flag_index {
         Some(indices) if indices.contains(',') => {
-            return match args.rconfig().indexed()? {
-                None => {
-                    let rconf = args.rconfig();
+            return {
+                let rconf = args.rconfig();
 
-                    if let Some(offset) = args.flag_byte_offset {
-                        let inner = rconf.io_reader_for_random_access()?;
-                        let mut rdr = rconf.csv_reader_from_reader(inner);
+                if let Some(offset) = args.flag_byte_offset {
+                    let inner = rconf.io_reader_for_random_access()?;
+                    let mut rdr = rconf.csv_reader_from_reader(inner);
 
-                        let mut pos = csv::Position::new();
-                        pos.set_byte(offset as u64);
+                    let mut pos = csv::Position::new();
+                    pos.set_byte(offset as u64);
 
-                        rdr.seek_raw(SeekFrom::Start(offset as u64), pos)?;
+                    rdr.seek_raw(SeekFrom::Start(offset as u64), pos)?;
 
-                        args.no_index_plural(rdr)
-                    } else {
-                        let rdr = rconf.reader()?;
-                        args.no_index_plural(rdr)
-                    }
+                    args.no_index_plural(rdr)
+                } else {
+                    let rdr = rconf.reader()?;
+                    args.no_index_plural(rdr)
                 }
-                Some(idx) => args.with_index_plural(idx),
             };
         }
         _ => (),
     };
 
-    match args.rconfig().indexed()? {
-        None => {
-            let rconf = args.rconfig();
+    let rconf = args.rconfig();
 
-            if let Some(offset) = args.flag_byte_offset {
-                let inner = rconf.io_reader_for_random_access()?;
-                let mut rdr = rconf.csv_reader_from_reader(inner);
+    if let Some(offset) = args.flag_byte_offset {
+        let inner = rconf.io_reader_for_random_access()?;
+        let mut rdr = rconf.csv_reader_from_reader(inner);
 
-                let mut pos = csv::Position::new();
-                pos.set_byte(offset as u64);
+        let mut pos = csv::Position::new();
+        pos.set_byte(offset as u64);
 
-                rdr.seek_raw(SeekFrom::Start(offset as u64), pos)?;
+        rdr.seek_raw(SeekFrom::Start(offset as u64), pos)?;
 
-                args.no_index(rdr)
-            } else {
-                let rdr = rconf.reader()?;
-                args.no_index(rdr)
-            }
-        }
-        Some(idx) => args.with_index(idx),
+        args.no_index(rdr)
+    } else {
+        let rdr = rconf.reader()?;
+        args.no_index(rdr)
     }
 }
 
@@ -158,26 +143,6 @@ impl Args {
         Ok(wtr.flush()?)
     }
 
-    fn with_index(&self, mut idx: Indexed<fs::File, fs::File>) -> CliResult<()> {
-        if self.flag_byte_offset.is_some() {
-            Err("-B/--byte-offset is pointless with indexed files!")?;
-        }
-
-        let mut wtr = self.wconfig().writer()?;
-        self.rconfig().write_headers(&mut *idx, &mut wtr)?;
-
-        let (start, end) = self.range()?;
-        if end - start == 0 {
-            return Ok(());
-        }
-        idx.seek(start as u64)?;
-        for r in idx.byte_records().take(end - start) {
-            wtr.write_byte_record(&r?)?;
-        }
-        wtr.flush()?;
-        Ok(())
-    }
-
     fn no_index_plural<R: Read>(&self, mut rdr: csv::Reader<R>) -> CliResult<()> {
         let mut wtr = self.wconfig().writer()?;
         self.rconfig().write_headers(&mut rdr, &mut wtr)?;
@@ -196,24 +161,6 @@ impl Args {
 
             if &i > indices.last().unwrap() {
                 break;
-            }
-        }
-
-        Ok(wtr.flush()?)
-    }
-
-    fn with_index_plural(&self, mut idx: Indexed<fs::File, fs::File>) -> CliResult<()> {
-        let mut wtr = self.wconfig().writer()?;
-        self.rconfig().write_headers(&mut *idx, &mut wtr)?;
-
-        for index in self.plural_indices()? {
-            idx.seek(index as u64)?;
-
-            match idx.byte_records().next() {
-                None => break,
-                Some(record) => {
-                    wtr.write_byte_record(&record?)?;
-                }
             }
         }